// src/types/logisticsRow.ts
import { OrderStatus, ProductCategory } from './orderStatus';

export interface LogisticsRow {
  id: number;
  orderDate: string;
  itemCode: string;
  itemName: string;
  qty: number;
  uom: string;
  supplierCode: string;
  supplierName: string;
  exporterName: string;
  brandName: string;
  orderedQty: number;
  unitPrice: number;
  currency: string;
  totalAmount: number;
  freightCostUSD: number;
  proformaNo: string;
  proformaDate: string;
  productionDeadline: string;
  expirationDate: string;
  docProductionDate: string;
  docExpirationDate: string;
  prepaymentPct: number;
  prepaymentTerms: string;
  prepaymentDays: number;
  finalPaymentPct: number;
  finalPaymentTerms: string;
  finalPaymentDays: number;
  prepaymentAmount: number;
  prepaymentDate: string;
  prepaymentConfirmDate: string;
  interimAmount: number;
  interimDate: string;
  warehouseEntryDate: string;
  containerPlannedLoadingDate: string;
  containerActualLoadingDate: string;
  exportCountry: string;
  dispatchInstructions: string;
  cargoPhotosReceived: string;
  transportMode: string;
  containerNo: string;
  plannedLoadingDate2: string;
  departureDateETD: string;
  currentDailyStatus: string;
  istanbulArrivalDate: string;
  istanbulDepartureDate: string;
  potiPlannedEta: string;
  potiActualArrival: string;
  containerPickupDate: string;
  containerReturnDate: string;
  freeDaysCount: number;
  secondaryDocsDate: string;
  courierCompany: string;
  courierTrackingNo: string;
  originalDocsDate: string;
  docsHandoverDate: string;
  invoiceNo: string;
  invoiceDate: string;
  invoicedQty: number;
  invoicedAmount: number;
  retentionAmount: number;
  finalPaymentAmount: number;
  plannedFinalPaymentDate: string;
  actualFinalPaymentDate: string;
  finalPaymentConfirmDate: string;
  plannedCustomsEntryDate: string;
  customsDeclarationNo: string;
  customsRegDate: string;
  customsReleaseDate: string;
  customsPaymentDate: string;
  freightCarrier: string;
  hsCode: string;
  transportExpense: number;
  status: OrderStatus;
  category: ProductCategory;
}